更新项目接口

+++ src/main/java/com/sw/chefubao/controller/CarTypeController.java
@@ -1,9 +1,12 @@
 package com.sw.chefubao.controller;
 
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.sw.chefubao.common.R;
 import com.sw.chefubao.entity.CarType;
+import com.sw.chefubao.entity.UserCar;
 import com.sw.chefubao.service.CarTypeService;
+import com.sw.chefubao.service.UserCarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,7 +20,8 @@ import java.util.List;
 public class CarTypeController {
     @Autowired
     private CarTypeService carTypeService;
-
+    @Autowired
+    private UserCarService userCarService;
 
     /**
      * 查询列表
@@ -37,7 +41,7 @@ public class CarTypeController {
      */
     @PostMapping("/save")
     public R save(@RequestBody CarType carType) {
-        boolean save = carTypeService.save(carType);
+        boolean save = carTypeService.saveOrUpdate(carType);
         if (!save) {
             return R.SAVE_ERROR;
         }
@@ -53,10 +57,22 @@ public class CarTypeController {
      */
     @PostMapping("/delete")
     public R delete(@RequestParam("id") Integer id) {
-        boolean b = carTypeService.removeById(id);
-        if (b) {
-            return R.DELETE_SUCCESS;
+        UserCar userCar = new UserCar();
+        userCar.setCarTypeId(id);
+        QueryWrapper<UserCar> userCarQueryWrapper = new QueryWrapper<>(userCar);
+        List<UserCar> list = userCarService.list(userCarQueryWrapper);
+        if (list.size() == 0) {
+            boolean b = carTypeService.removeById(id);
+            if (b) {
+                return R.DELETE_SUCCESS;
+            }
+        } else {
+            R r = new R<>();
+            r.setCode(506);
+            r.setMessage("用户已有该类型的车辆，禁止删除");
+            return r;
         }
+
         return R.DELETE_ERROR;
     }
 }
